Added the OpenJobs Database
I build the OpenJobs database to store the ticket information. I used the entity framework instead of the traditional MySQL because its easier to write and well integrated + supported by ASP.NET. All that was added was the database, model C# file, DbContext C# file, and the connection string to the web.config file.