Added controllers & views
I added the views called ServiceRequestForm and ServiceRequestDetails. The ServiceRequestController helps with creating these views, as well as routing the openJobModel object from view to view. The controller named TechnicianInterfaceController was added. This will have to edited to determine how much control we want to give Technicians over the database data. Also there is lot of test data in the database that should be deleted. The database may have to be migrated again (if we alter it) so I figured we could delete the data then.