Added a security library
I am not sure if this was necessary since Microsoft might have included this, but I added it anyway. I also attempted to encode the username/password for the login page. XSS attacks are blocked but I am not sure if they have always been or if the encoding helped prevent them. The encoding was done on the AccountController.